Add files via upload
Since the files are poorly named... "lookingforcomplex.py" is the code used for the 3d surface plots I have shared frequently and used for my undergrad thesis, "sinosf.py" does the cross sections that we recently settled on agreement with, "sinosfupdate.py" is my pole finder code, "delta_apply_residue_loopg.py" is the code the produces the scalar form factor as a function of g, "physfunc.py" is the collection of important functions I add to and pull from. I believe all the editions in their current form are ready to run as long as all files are in the same folder but I have not looked at much of this in awhile and I do not fully remember them all.